Changed url for IO from SmashTheStack to netgarage
The IO wargame is no longer hosted at smashthestack, and is now located at the new URL given in this commit (io.netgarage.org). While playing that wargame, I spoke on IRC with its author (bla) and he requested that the reference be updated, as he has parted ways with smashthestack. For posterity, I kept the mention of smashthestack in brackets after the link, to be clear for players who had completed the challenge while it was at its old host.